+++ lib/tlGL/Render.cpp
@@ -292,7 +292,7 @@ namespace tl
         {
             std::vector<std::shared_ptr<Texture> > out;
             const auto i = std::find_if(
-                _cache.begin(), 
+                _cache.begin(),
                 _cache.end(),
                 [info, imageFilters](const TextureData& value)
                 {
@@ -373,7 +373,12 @@ namespace tl
             out->_init(context);
             return out;
         }
-        
+
+        std::shared_ptr<Shader> Render::getShader( const std::string& name )
+        {
+            return _p->shaders[name];
+        }
+
         void Render::setTextureCacheSize(size_t value)
         {
             _p->textureCache.setSize(value);
@@ -836,5 +841,6 @@ namespace tl
 
         void Render::end()
         {}
+
     }
 }
