+++ lib/tlGL/Render.h
@@ -10,6 +10,8 @@ namespace tl
 {
     namespace gl
     {
+        class Shader;
+
         //! OpenGL renderer.
         class Render : public timeline::IRender
         {
@@ -25,6 +27,7 @@ namespace tl
             //! Create a new renderer.
             static std::shared_ptr<Render> create(const std::shared_ptr<system::Context>&);
 
+            std::shared_ptr<Shader> getShader( std::string name );
             void setTextureCacheSize(size_t) override;
             void setColorConfig(const timeline::ColorConfigOptions&) override;
             void setLUT(const timeline::LUTOptions&) override;
@@ -51,7 +54,8 @@ namespace tl
                 const std::vector<timeline::ImageOptions>& = {},
                 const std::vector<timeline::DisplayOptions>& = {},
                 const timeline::CompareOptions& = timeline::CompareOptions()) override;
-
+            std::shared_ptr<Shader> getShader( const std::string& );
+            
         private:
             void _drawVideo(
                 const timeline::VideoData&,
